Rename: clean up file names
파일 이름 정리
