+++ app/src/main/AndroidManifest.xml
@@ -90,10 +90,10 @@
             android:value="atguginznv" />
 
         <activity
-            android:name=".MainActivity2"
+            android:name=".LoginActivity"
             android:exported="true"></activity>
         <activity
-            android:name=".MainActivty"
+            android:name=".MainActivity"
             android:exported="true" />
         <activity android:name=".RegisterActivity" />
         <activity android:name=".FindActivity" />
