+++ app/src/main/res/layout/activity_login.xml
@@ -4,36 +4,37 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@color/background"
     android:orientation="vertical"
-    >
+    android:padding="20dp">
 
     <ImageView
         android:layout_width="match_parent"
         android:layout_height="170dp"
         android:layout_marginTop="30dp"
-        android:src="@drawable/logo_aob" />
+        android:src="@drawable/dog" />
 
     <TextView
-        android:layout_width="350dp"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
         android:layout_marginTop="30dp"
         android:fontFamily="@font/nanum_barun_gothic"
-        android:text="로그인하자~"
-        android:textColor="@color/colorAccent"
+        android:text="로그인"
+        android:textColor="@color/brown"
         android:textSize="40sp" />
 
     <EditText
         android:id="@+id/login_id"
-        android:layout_width="350dp"
+        android:layout_width="300dp"
         android:layout_height="50dp"
         android:layout_gravity="center"
         android:layout_marginTop="20dp"
         android:background="@drawable/style_rounded_login"
-        android:fontFamily="@font/nanum_barun_gothic_light"
-        android:inputType="textVisiblePassword"
         android:digits="@string/text_id"
+        android:fontFamily="@font/nanum_barun_gothic_light"
         android:hint="아이디를 입력해주세요."
+        android:inputType="textVisiblePassword"
         android:paddingLeft="17dp"
         android:textColor="@color/black"
         android:textColorHint="@color/black"
@@ -41,20 +42,20 @@
 
     <TextView
         android:id="@+id/login_id_check"
-        android:layout_width="330dp"
+        android:layout_width="300dp"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
-        android:fontFamily="@font/nanum_square"
-        android:text="아이디를 입력해 주세요."
         android:layout_marginTop="6dp"
+        android:fontFamily="@font/nanum_square"
+        android:maxLength="15"
+        android:text="아이디를 입력해 주세요."
         android:textColor="#FF0000"
         android:textSize="12sp"
-        android:maxLength="15"
-        android:visibility="gone"/>
+        android:visibility="gone" />
 
     <EditText
         android:id="@+id/login_pw"
-        android:layout_width="350dp"
+        android:layout_width="300dp"
         android:layout_height="50dp"
         android:layout_gravity="center"
         android:layout_marginTop="15dp"
@@ -62,57 +63,58 @@
         android:fontFamily="@font/nanum_barun_gothic_light"
         android:hint="비밀번호를 입력해주세요."
         android:inputType="textPassword"
+        android:maxLength="16"
         android:paddingLeft="17dp"
         android:textColor="@color/black"
         android:textColorHint="@color/black"
-        android:maxLength="16"
         android:textSize="15sp" />
 
     <TextView
         android:id="@+id/login_pw_check"
-        android:layout_width="330dp"
+        android:layout_width="300dp"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
+        android:layout_marginTop="6dp"
         android:fontFamily="@font/nanum_square"
         android:text="비밀번호를 입력해주세요."
-        android:layout_marginTop="6dp"
         android:textColor="#FF0000"
         android:textSize="12sp"
-        android:visibility="gone"/>
+        android:visibility="gone" />
 
     <TextView
         android:id="@+id/login_check"
-        android:layout_width="330dp"
+        android:layout_width="300dp"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
+        android:layout_marginTop="6dp"
         android:fontFamily="@font/nanum_square"
         android:text="아이디, 비밀번호를 다시 확인해주세요."
-        android:layout_marginTop="6dp"
         android:textColor="#FF0000"
         android:textSize="12sp"
-        android:visibility="gone"/>
+        android:visibility="gone" />
 
     <LinearLayout
-        android:layout_width="350dp"
+        android:layout_width="300dp"
         android:layout_height="45dp"
         android:layout_gravity="center"
+        android:layout_marginTop="5dp"
         android:orientation="horizontal">
 
         <CheckBox
             android:id="@+id/login_stay_cb"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:button="@null"
             android:layout_marginBottom="1dp"
-            android:background="@drawable/style_checkbox"/>
+            android:background="@drawable/style_checkbox"
+            android:button="@null" />
 
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="match_parent"
-            android:gravity="center"
-            android:fontFamily="@font/nanum_square"
-            android:text="로그인 상태 유지"
             android:layout_marginLeft="5dp"
+            android:fontFamily="@font/nanum_square"
+            android:gravity="center"
+            android:text="로그인 상태 유지"
             android:textColor="@color/gray"
             android:textSize="14sp" />
 
@@ -121,22 +123,16 @@
             android:layout_height="0dp"
             android:layout_weight="1" />
 
-        <Button
-            android:id="@+id/login_find"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:background="@color/transparent"
-            android:fontFamily="@font/nanum_square"
-            android:text="아이디 / 비밀번호 찾기"
-            android:textSize="14sp"
-            android:textColor="@color/gray" />
-
     </LinearLayout>
+
+
     <Button
         android:id="@+id/login_button"
-        android:layout_width="350dp"
+        android:layout_width="150dp"
         android:layout_height="50dp"
         android:layout_gravity="center"
+        android:layout_marginTop="30dp"
+        android:layout_marginBottom="10dp"
         android:background="@drawable/style_rounded_postive_button"
         android:fontFamily="@font/nanum_square_eb"
         android:text="로그인"
@@ -144,30 +140,35 @@
         android:textStyle="bold" />
 
     <LinearLayout
-        android:layout_width="350dp"
+        android:layout_width="300dp"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
         android:gravity="right"
-        android:orientation="horizontal">
+        android:orientation="horizontal"
+        android:weightSum="3">
 
-        <TextView
+        <Button
+            android:id="@+id/login_find"
             android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
+            android:layout_height="match_parent"
+            android:layout_gravity="center"
+            android:layout_weight="1"
             android:background="@color/transparent"
             android:fontFamily="@font/nanum_square"
-            android:text="처음 방문하셨나요?"
-            android:textSize="15sp"
-            android:textColor="@color/gray" />
+            android:text="아이디 / 비밀번호 찾기"
+            android:textColor="@color/gray"
+            android:textSize="14sp" />
 
         <Button
             android:id="@+id/login_register"
-            android:layout_width="75dp"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_weight="2"
             android:background="@color/transparent"
             android:fontFamily="@font/nanum_square_eb"
             android:text="회원가입"
-            android:textSize="15sp"
-            android:textColor="@color/green" />
+            android:textColor="@color/brown"
+            android:textSize="15sp" />
 
     </LinearLayout>
